pointer to arg added to timer (handling tcp/tcb timeouts), tcp: listen state, sending packet (without data)
git-svn-id: https://avr-net.googlecode.com/svn/trunk@22 a018619c-9330-2494-4511-e33c901fd56f

+++ src/net/net.c
@@ -16,7 +16,7 @@ uint16_t 	hton16(uint16_t h)
 	return HTON16(h);
 }
 
-uint32_t 	hton32(uint16_t h)
+uint32_t 	hton32(uint32_t h)
 {
 	return HTON32(h);
 }
